Convert pnValue to float in NewthonRapson
pnValue was previously parsed as an integer in NewthonRapson.tsx. This commit updates the code to parse pnValue as a float instead, as it allows for more precision in calculations.

+++ src/Views/NewthonRapson.tsx
@@ -40,7 +40,7 @@ export default function NewthonRapson() {
             setDerivadaLaTex("f^{\\prime}\\left(x\\right)=" + derivadaNode.toTex());
             setOriginal(originalNode.compile());
             setDerivada(derivadaNode.compile());
-            let pn = Number.parseInt(pnValue);
+            let pn = Number.parseFloat(pnValue);
             let firstElement: Values = {
                 error: undefined,
                 pn,
